Shift and combine files to better adhere to MVC pattern
