--- life-map/routes/auth/logout.js
@@ -1,10 +1,0 @@
-const express = require('express')
-
-const router = express.Router()
-
-router.post('/', (req, res) => {
-  req.logOut()
-  res.redirect('/login')
-})
-
-module.exports = router;
