+++ life-map/server.js
@@ -57,8 +57,8 @@ app.post('/login', passport.authenticate('local', {
 
 
 // Routes
-app.use('/', require('./route_includes/auth_routes.js'));
-app.use('/homepage', require('./route_includes/homepage_routes.js'));
+app.use('/', require('./routes/auth.js'));
+app.use('/homepage', require('./routes/homepage.js'));
 
 app.listen(PORT, console.log(`Server running on port ${PORT}`));
